# Update docker images for performance and implement async bot commands