--- a/reality-ezpz.sh
+++ b/reality-ezpz.sh
@@ -972,7 +972,7 @@ function generate_tgbot_dockerfile {
 FROM ${image[python]}
 WORKDIR ${config_path}/tgbot
 RUN apk add --no-cache docker-cli-compose curl bash newt libqrencode-tools sudo openssl jq zip unzip
-RUN pip install --no-cache-dir python-telegram-bot==22.0 qrcode[pil]==8.1
+RUN pip install --no-cache-dir python-telegram-bot[callback-data]==22.0 qrcode[pil]==8.1
 CMD [ "python", "./tgbot.py" ]
 EOF
 }