--- a/tgbot.py
+++ b/tgbot.py
@@ -1,7 +1,7 @@
+import asyncio
 import io
 import os
 import re
-import subprocess
 
 import qrcode
 from telegram import InlineKeyboardButton, InlineKeyboardMarkup, Update
@@ -20,34 +20,94 @@ username_regex = re.compile("^[a-zA-Z0-9]+$")
 command = "bash <(curl -sL https://raw.githubusercontent.com/aleskxyz/reality-ezpz/master/reality-ezpz.sh) "
 
 
-def get_users_ezpz():
-    local_command = command + "--list-users"
-    return run_command(local_command).split("\n")[:-1]
+class ScriptExecutionError(Exception):
+    """Custom exception for errors when running the external script."""
+
+    def __init__(self, message, stdout, stderr, returncode):
+        super().__init__(message)
+        self.stdout = stdout
+        self.stderr = stderr
+        self.returncode = returncode
+
+    def __str__(self):
+        return f"{super().__str__()} (Exit code: {self.returncode})\nStderr: {self.stderr}\nStdout: {self.stdout}"
 
 
-def get_config_ezpz(username):
-    local_command = command + f"--show-user {username} | grep -E '://|^\\{{\"dns\"'"
-    return run_command(local_command).split("\n")[:-1]
-
-
-def delete_user_ezpz(username):
-    local_command = command + f"--delete-user {username}"
-    run_command(local_command)
-    return
-
-
-def add_user_ezpz(username):
-    local_command = command + f"--add-user {username}"
-    run_command(local_command)
-    return
-
-
-def run_command(command):
-    process = subprocess.Popen(
-        ["/bin/bash", "-c", command], stdout=subprocess.PIPE, stderr=subprocess.PIPE
+async def async_run_script_command(cmd: str) -> str:
+    """
+    Run a shell command asynchronously and return stdout.
+    Raises ScriptExecutionError if the command fails.
+    """
+    # Use asyncio.create_subprocess_exec for async subprocess
+    # Need shell=True because the command string uses shell features like <()
+    process = await asyncio.create_subprocess_exec(
+        "/bin/bash",
+        "-c",
+        cmd,
+        stdout=asyncio.subprocess.PIPE,
+        stderr=asyncio.subprocess.PIPE,
     )
-    output, _ = process.communicate()
-    return output.decode()
+
+    stdout, stderr = await process.communicate()
+
+    stdout_str = stdout.decode().strip()
+    stderr_str = stderr.decode().strip()
+
+    if process.returncode != 0:
+        raise ScriptExecutionError(
+            f"Script command failed: {cmd}",
+            stdout=stdout_str,
+            stderr=stderr_str,
+            returncode=process.returncode,
+        )
+
+    return stdout_str
+
+
+async def get_users_ezpz():
+    local_command = command + "--list-users"
+    try:
+        output = await async_run_script_command(local_command)
+        users = output.split("\n")
+        if users and users[-1] == "":
+            users.pop()
+        return users
+    except ScriptExecutionError as e:
+        print(f"Error listing users: {e}")
+        return []
+
+
+async def get_config_ezpz(username: str):
+    local_command = command + f"--show-user {username} | grep -E '://|^\\{{\"dns\"'"
+    try:
+        output = await async_run_script_command(local_command)
+        configs = output.split("\n")
+        if configs and configs[-1] == "":
+            configs.pop()
+        return configs
+    except ScriptExecutionError as e:
+        print(f"Error getting config for {username}: {e}")
+        return []
+
+
+async def delete_user_ezpz(username: str):
+    local_command = command + f"--delete-user {username}"
+    try:
+        await async_run_script_command(local_command)
+        return True
+    except ScriptExecutionError as e:
+        print(f"Error deleting user {username}: {e}")
+        return False
+
+
+async def add_user_ezpz(username: str):
+    local_command = command + f"--add-user {username}"
+    try:
+        await async_run_script_command(local_command)
+        return True
+    except ScriptExecutionError as e:
+        print(f"Error adding user {username}: {e}")
+        return False
 
 
 def restricted(func):
@@ -60,12 +120,12 @@ def restricted(func):
         elif update.callback_query and update.callback_query.message:
             username = update.callback_query.message.chat.username
         admin_list = admin.split(",")
-        if username in admin_list:
+        if username and username in admin_list:
             return await func(update, context, *args, **kwargs)
         else:
             await context.bot.send_message(
                 chat_id=update.effective_chat.id,
-                text="You are not authorized to use this bot.",
+                text="You are not authorized to use this bot. Your username is not in the admin list.",
             )
 
     return wrapped
@@ -87,10 +147,21 @@ async def start(update: Update, context: ContextTypes.DEFAULT_TYPE):
 
 @restricted
 async def users_list(
-    update: Update, context: ContextTypes.DEFAULT_TYPE, text, callback
+    update: Update, context: ContextTypes.DEFAULT_TYPE, text: str, callback: str
 ):
     keyboard = []
-    for user in get_users_ezpz():
+    users = await get_users_ezpz()
+    if not users:
+        await context.bot.send_message(
+            chat_id=update.effective_chat.id,
+            text="Could not retrieve user list or no users found.",
+            reply_markup=InlineKeyboardMarkup(
+                [[InlineKeyboardButton("Back", callback_data="start")]]
+            ),
+        )
+        return
+
+    for user in users:
         keyboard.append(
             [InlineKeyboardButton(user, callback_data=f"{callback}!{user}")]
         )
@@ -102,15 +173,26 @@ async def users_list(
 
 
 @restricted
-async def show_user(update: Update, context: ContextTypes.DEFAULT_TYPE, username):
+async def show_user(update: Update, context: ContextTypes.DEFAULT_TYPE, username: str):
     keyboard = [[InlineKeyboardButton("Back", callback_data="show_user")]]
     reply_markup = InlineKeyboardMarkup(keyboard)
+
     await context.bot.send_message(
         chat_id=update.effective_chat.id,
-        text=f'Config for "{username}":',
+        text=f'Fetching config for "{username}"...',
         parse_mode="HTML",
     )
-    config_list = get_config_ezpz(username)
+
+    config_list = await get_config_ezpz(username)
+
+    if not config_list:
+        await context.bot.send_message(
+            chat_id=update.effective_chat.id,
+            text=f'Could not retrieve config for "{username}". The user might not exist or there was a script error.',
+            reply_markup=reply_markup,
+        )
+        return
+
     ipv6_pattern = r'"server":"[0-9a-fA-F:]+"'
 
     for config in config_list:
@@ -119,24 +201,38 @@ async def show_user(update: Update, context: ContextTypes.DEFAULT_TYPE, username
         else:
             config_text = f"<pre>{config}</pre>"
 
-        qr_img = qrcode.make(config)
-        bio = io.BytesIO()
-        qr_img.save(bio, "PNG")
-        bio.seek(0)
+        try:
+            qr_img = qrcode.make(config)
+            bio = io.BytesIO()
+            qr_img.save(bio, "PNG")
+            bio.seek(0)
 
-        await context.bot.send_photo(
-            chat_id=update.effective_chat.id,
-            photo=bio,
-            caption=config_text,
-            parse_mode="HTML",
-            reply_markup=reply_markup,
-        )
+            await context.bot.send_photo(
+                chat_id=update.effective_chat.id,
+                photo=bio,
+                caption=config_text,
+                parse_mode="HTML",
+                reply_markup=reply_markup,
+            )
+        except Exception as e:
+            print(
+                f"Error generating QR or sending photo for config: {config}. Error: {e}"
+            )
+            await context.bot.send_message(
+                chat_id=update.effective_chat.id,
+                text=f"Could not generate QR code for the following config:\n{config_text}",
+                parse_mode="HTML",
+                reply_markup=reply_markup,
+            )
 
 
 @restricted
-async def delete_user(update: Update, context: ContextTypes.DEFAULT_TYPE, username):
+async def delete_user(
+    update: Update, context: ContextTypes.DEFAULT_TYPE, username: str
+):
     keyboard = []
-    if len(get_users_ezpz()) == 1:
+    users = await get_users_ezpz()
+    if len(users) == 1 and username in users:
         text = "You cannot delete the only user.\nAt least one user is needed.\nCreate a new user, then delete this one."
         keyboard.append([InlineKeyboardButton("Back", callback_data="start")])
         reply_markup = InlineKeyboardMarkup(keyboard)
@@ -168,12 +264,19 @@ async def add_user(update: Update, context: ContextTypes.DEFAULT_TYPE):
 
 
 @restricted
-async def approve_delete(update: Update, context: ContextTypes.DEFAULT_TYPE, username):
-    delete_user_ezpz(username)
-    text = f"User {username} has been deleted."
+async def approve_delete(
+    update: Update, context: ContextTypes.DEFAULT_TYPE, username: str
+):
+    success = await delete_user_ezpz(username)
     keyboard = []
     keyboard.append([InlineKeyboardButton("Back", callback_data="start")])
     reply_markup = InlineKeyboardMarkup(keyboard)
+
+    if success:
+        text = f"User {username} has been deleted."
+    else:
+        text = f"Failed to delete user {username}. Check bot logs for details."
+
     await context.bot.send_message(
         chat_id=update.effective_chat.id, text=text, reply_markup=reply_markup
     )
@@ -190,6 +293,14 @@ async def cancel(update: Update, context: ContextTypes.DEFAULT_TYPE):
 async def button(update: Update, context: ContextTypes.DEFAULT_TYPE):
     query = update.callback_query
     await query.answer()
+
+    # Delete the message with the inline keyboard to avoid stale buttons
+    if query.message:
+        try:
+            await query.message.delete()
+        except Exception as e:
+            print(f"Could not delete message: {e}")
+
     response = query.data.split("!")
     if len(response) == 1:
         if response[0] == "start":
@@ -204,56 +315,82 @@ async def button(update: Update, context: ContextTypes.DEFAULT_TYPE):
             await users_list(update, context, "Select user to delete:", "delete_user")
         elif response[0] == "add_user":
             await add_user(update, context)
+    if len(response) > 1:
+        action = response[0]
+        username = response[1]
+
+        if action == "show_user":
+            await show_user(update, context, username)
+        elif action == "delete_user":
+            await delete_user(update, context, username)
+        elif action == "approve_delete":
+            await approve_delete(update, context, username)
         else:
             await context.bot.send_message(
                 chat_id=update.effective_chat.id,
-                text="Button pressed: {}".format(response[0]),
+                text=f"Unknown action: {action}",
             )
-    if len(response) > 1:
-        if response[0] == "show_user":
-            await show_user(update, context, response[1])
-        if response[0] == "delete_user":
-            await delete_user(update, context, response[1])
-        if response[0] == "approve_delete":
-            await approve_delete(update, context, response[1])
 
 
 @restricted
 async def user_input(update: Update, context: ContextTypes.DEFAULT_TYPE):
     if "expected_input" in context.user_data:
         expected_input = context.user_data["expected_input"]
-        del context.user_data["expected_input"]
+
         if expected_input == "username":
-            username = update.message.text
-            if username in get_users_ezpz():
+            username = update.message.text.strip()
+
+            existing_users = await get_users_ezpz()
+            if username in existing_users:
                 await update.message.reply_text(
                     f'User "{username}" exists, try another username.'
                 )
                 await add_user(update, context)
                 return
+
             if not username_regex.match(username):
                 await update.message.reply_text(
-                    "Username can only contains A-Z, a-z and 0-9, try another username."
+                    "Username can only contain A-Z, a-z and 0-9, try another username."
                 )
                 await add_user(update, context)
                 return
-            add_user_ezpz(username)
-            await update.message.reply_text(f'User "{username}" is created.')
-            await show_user(update, context, username)
+
+            success = await add_user_ezpz(username)
+
+            if success:
+                await update.message.reply_text(f'User "{username}" is created.')
+                del context.user_data["expected_input"]
+                await show_user(update, context, username)
+            else:
+                await update.message.reply_text(
+                    f'Failed to create user "{username}". Check bot logs for details.'
+                )
+                del context.user_data["expected_input"]
+                await start(update, context)
+        else:
+            print(f"Warning: Unknown expected_input state: {expected_input}")
+            del context.user_data["expected_input"]
+            await update.message.reply_text(
+                "An unexpected state occurred. Returning to start."
+            )
+            await start(update, context)
 
 
 def main():
-    # Create the Application
-    application = Application.builder().token(token).build()
+    application = (
+        Application.builder().token(token).arbitrary_callback_data(True).build()
+    )
 
-    # Add handlers
     application.add_handler(CommandHandler("start", start))
-    application.add_handler(CallbackQueryHandler(button))
+    application.add_handler(CallbackQueryHandler(button, pattern=".*"))
     application.add_handler(MessageHandler(filters.TEXT & ~filters.COMMAND, user_input))
 
-    # Start the Bot - using the non-blocking approach
+    print("Bot started. Waiting for messages...")
     application.run_polling(allowed_updates=Update.ALL_TYPES)
 
 
 if __name__ == "__main__":
-    main()
+    try:
+        main()
+    except Exception as ex:
+        print(f"Bot failed to start: {ex}")