Updates (`reality-ezpz.sh`):
- Updated the Docker image versions
- Observed significant performance improvements with these newer versions:
  - sing-box speed improved from ~150 Mbps to ~600 Mbps on a 1 Gbps connection.
  - xray speed improved from ~160 Mbps to ~210 Mbps on a 1 Gbps connection.

(Performance tested on an Ampere Altra 4-core system.)

Bot Improvements (`tgbot.py`):
- Implemented asynchronous execution for all calls to the reality-ezpz.sh script
  using `asyncio`. This prevents the bot from freezing while waiting for script
  commands, ensuring a smoother and more responsive user experience.
- Added robust error handling for script executions. The bot now catches script
  failures, logs detailed errors server-side, and provides informative feedback
  to the user in Telegram.