@ovyan 2025-08-24
@aleskxyz can we merge this PR or at least update images?
<img width="1184" height="526" alt="image" src="https://github.com/user-attachments/assets/24ac3096-f262-470b-96a1-d9ef562b3b21" />
I am getting severe memory leaks caused by gzxhwq/sing-box:1.8.14 image on Debian 12