@ovyan 2025-08-25
FYI: i created a backup, then uninstalled ezpz, installed using my custom script with modified images and then successfully restored from backup. Though, i had to keep python3.11 instead of python 3.13 because it did not work out of the box (i did not have these async changes in my code)