@hingston 2025-10-06
@aleskxyz Could you please take a look at this PR when you have a moment? I've had many people using this version already.

@xiaat, @ovyan For anyone who wants to test this version, you can use the following command:

```bash
bash <(curl -sL https://raw.githubusercontent.com/hingston/reality-ezpz/refs/heads/test/reality-ezpz.sh) -m
```